Resolve CI g++ and clang issues

+++ src/process_line.h
@@ -47,7 +47,7 @@ class PostProcessSingleComponent final : public ProcessLine
 public:
     PostProcessSingleComponent(void* rawData, const JlsParameters& params, size_t bytesPerPixel) noexcept :
         rawData_{static_cast<uint8_t*>(rawData)},
-        bytesPerPixel_{static_cast<size_t>(bytesPerPixel)},
+        bytesPerPixel_{bytesPerPixel},
         bytesPerLine_{static_cast<size_t>(params.stride)}
     {
     }
